backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
firebase-admin>=6.5.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
openai>=1.30.0
python-dotenv>=1.0.1
